wifi: Set Vendor lib dependency on library under vendor partition
Currently supplicant vendor library is marked dependent on library
which is generated under system partition. When system image is
replaced with GSI image, vendor library would not be able to locate
library which was generated in system.

The dependency is resolved by making vendor lib to use
libraries from vendor partition.

CRs-Fixed: 2261557
Change-Id: I3a9ff021650c2a5d1fd4096dee47a15c5c3a8dc6